+++ src/r1/scr11c.c
@@ -482,7 +482,7 @@ void scrollc_h(Sint32 lD4, Sint32 flagc) {
 
 scrc_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) scrflagc.b.h |= flagc;
+    if (lD0.l < 0) scrflagc.b.h |= flagc;
     else scrflagc.b.h |= flagc * 2;
   }
 }
