+++ src/r3/scr32d.c
@@ -223,8 +223,8 @@ scrflaga.w = scrflagb.w = scrflagc.w = scrflagz.w = 0;
   vscroll.w.l = scrb_v_posit.w.h;
 
 scrollz_h(scra_hz << 4, 64);
-  scrollc_h((long int)(scra_hz << 3) * 3, 16);
-  scrollb_h((long int)(scra_hz << 2) * 3, 4);
+  scrollc_h((scra_hz << 3) * 3, 16);
+  scrollb_h((scra_hz << 2) * 3, 4);
 
   lD0.w.l = 536;
   wD1 = 1312;
@@ -258,7 +258,7 @@ pHscrWk = (int_union*)hscrollwork;
   }
   wD1 = 0;
   for ( ; i < 80; ++i) {
-    lD2.l = ((long int)(Sint32)wD1 << 8) + 32768;
+    lD2.l = ((Sint32)wD1 << 8) + 32768;
     pHscrWk->l += lD2.l;
     ++pHscrWk;
     ++wD1;
@@ -764,7 +764,7 @@ void scrollb_hv(int_union lD4, int_union lD5) {
 
     scrb_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagb.b.h |= 4;
     }
     else {
@@ -781,7 +781,7 @@ lD0.l = lD3.l = scrb_v_posit.l;
 
     scrb_v_count ^= 16;
     lD0.l -= lD3.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagb.b.h |= 1;
     }
     else {
@@ -828,7 +828,7 @@ void scrollb_h(Sint32 lD4, Sint32 flagb) {
   if (!lD1.b.b4) {
     scrb_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagb.b.h |= flagb;
     }
     else {
@@ -856,7 +856,7 @@ void scrollc_h(Sint32 lD4, Sint32 flagc) {
   if (!lD1.b.b4) {
     scrc_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagc.b.h |= flagc;
     }
     else {
@@ -884,7 +884,7 @@ void scrollz_h(Sint32 lD4, Sint32 flagz) {
   if (!lD1.b.b4) {
     scrz_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagz.b.h |= flagz;
     }
     else {
