+++ src/r4/scr42c.c
@@ -200,7 +200,7 @@ scrflaga.w = scrflagb.w = scrflagc.w = scrflagz.w = 0;
   vscroll.w.h = scra_v_posit.w.h;
   vscroll.w.l = scrb_v_posit.w.h;
 
-temp1.l = (long int)(scra_hz << 2) * 3;
+temp1.l = (scra_hz << 2) * 3;
   temp2.l = scra_vz * 128;
   scrollb_hv(&temp1, &temp2);
 
@@ -580,7 +580,7 @@ void scrollb_hv(int_union* lD4, int_union* lD5) {
 
     scrb_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagb.b.h |= 4;
     }
     else {
@@ -591,14 +591,14 @@ void scrollb_hv(int_union* lD4, int_union* lD5) {
 
 lD0.l = lD3.l = scrb_v_posit.l;
   lD0.l += lD5->l;
-  if ((long int)lD0.l < 0) lD0.l = 0;
+  if (lD0.l < 0) lD0.l = 0;
   scrb_v_posit.l = lD0.l;
 
   if (((Uint8)(scrb_v_posit.w.h & 16) ^ scrb_v_count) == 0) {
 
     scrb_v_count ^= 16;
     lD0.l -= lD3.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagb.b.h |= 1;
     }
     else {
@@ -645,7 +645,7 @@ void scrollb_h(Sint32 lD4, Sint32 flagb) {
   if (!lD1.b.b4) {
     scrb_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagb.b.h |= flagb;
     }
     else {
@@ -673,7 +673,7 @@ void scrollc_h(Sint32 lD4, Sint32 flagc) {
   if (!lD1.b.b4) {
     scrc_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagc.b.h |= flagc;
     }
     else {
@@ -701,7 +701,7 @@ void scrollz_h(Sint32 lD4, Sint32 flagz) {
   if (!lD1.b.b4) {
     scrz_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagz.b.h |= flagz;
     }
     else {
