+++ src/r4/scr43c.c
@@ -194,7 +194,7 @@ scrflaga.w = scrflagb.w = scrflagc.w = scrflagz.w = 0;
   vscroll.w.h = scra_v_posit.w.h;
   vscroll.w.l = scrb_v_posit.w.h;
 
-temp1.l = (long int)(scra_hz << 2) * 3;
+temp1.l = (scra_hz << 2) * 3;
   temp2.l = scra_vz << 7;
   scrollb_hv(&temp1, &temp2);
 
@@ -574,7 +574,7 @@ void scrollb_hv(int_union* lD4, int_union* lD5) {
 
     scrb_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagb.b.h |= 4;
     }
     else {
@@ -585,14 +585,14 @@ void scrollb_hv(int_union* lD4, int_union* lD5) {
 
 lD0.l = lD3.l = scrb_v_posit.l;
   lD0.l += lD5->l;
-  if ((long int)lD0.l < 0) lD0.l = 0;
+  if (lD0.l < 0) lD0.l = 0;
   scrb_v_posit.l = lD0.l;
 
   if (((Uint8)(scrb_v_posit.w.h & 16) ^ scrb_v_count) == 0) {
 
     scrb_v_count ^= 16;
     lD0.l -= lD3.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagb.b.h |= 1;
     }
     else {
@@ -639,7 +639,7 @@ void scrollb_h(Sint32 lD4, Sint32 flagb) {
   if (!lD1.b.b4) {
     scrb_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagb.b.h |= flagb;
     }
     else {
@@ -667,7 +667,7 @@ void scrollc_h(Sint32 lD4, Sint32 flagc) {
   if (!lD1.b.b4) {
     scrc_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagc.b.h |= flagc;
     }
     else {
@@ -695,7 +695,7 @@ void scrollz_h(Sint32 lD4, Sint32 flagz) {
   if (!lD1.b.b4) {
     scrz_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagz.b.h |= flagz;
     }
     else {
